requirements/base.txt
Django==1.11.4
django-allauth==0.33.0
django-braces==1.11.0
django-environ==0.4.4
django-model-utils==3.0.0
django-widget-tweaks==1.4.1
model-mommy==1.4.0
psycopg2==2.7.3.1
